feat: update rental process

+++ src/routes/pages/IndividualItem.tsx
@@ -6,6 +6,7 @@ import {
   ImageItem,
   IndividualItemResult,
   Loading,
+  RentRental,
   ReplaceRental,
   TransferItem,
 } from '../../components';
@@ -18,49 +19,44 @@ const IndividualItem: FC = () => {
   const { id } = useParams<{ id: string }>();
   // get individual item result
   const result: IndividualItemResponse | ErrorResponse | Pending = useFetchIndividualItem(id);
+  //早期リターン
+  if (typeof id === 'undefined') {
+    // 発生しないはず
+    return <h2>Unexpected Error</h2>;
+  }
   return (
     <>
-      {typeof id === 'undefined' ? (
-        // 発生しないはず
-        <h2>Unexpected Error</h2>
+      {result === 'pending' ? (
+        // 処理中
+        <Loading />
+      ) : 'code' in result && 'message' in result ? (
+        // fetchに失敗
+        <ErrorResult result={result} />
       ) : (
+        // fetch成功
         <>
-          {result === 'pending' ? (
-            // 処理中
-            <Loading />
-          ) : 'code' in result && 'message' in result ? (
-            // fetchに失敗
-            <ErrorResult result={result} />
+          <IndividualItemResult result={result} />
+          {/*Update*/}
+          {result.is_rent ? (
+            <Link to={``}>貸出中の物品は編集できません。</Link>
           ) : (
-            // fetch成功
-            <>
-              <IndividualItemResult result={result} />
-              {/*Update*/}
-              {result.is_rent ? (
-                <Link to={``}>貸出中の物品は編集できません。</Link>
-              ) : (
-                <Link to={`/item/${id}/update`}>更新リンク</Link>
-              )}
-              <br />
-              {/*Image*/}
-              <ImageItem id={id} isHidden={result.is_rent ? true : false} />
-              {/*Rental*/}
-              {result.is_rent ? (
-                <>
-                  <Link to={`/rental/${id}/update`}>貸し出し情報の更新</Link>
-                  <ReplaceRental id={id} />
-                </>
-              ) : (
-                <Link to={`/rental/${id}/rent`}>貸し出し</Link>
-              )}
-              <br />
-              {/*Transfer*/}
-              <TransferItem id={id} isHidden={result.id === 1 || result.is_rent ? true : false} />
-              <br />
-              {/*Delete*/}
-              <DeleteItem id={id} isHidden={result.id === 1 || result.is_rent ? true : false} />
-            </>
+            <Link to={`/item/${id}/update`}>更新リンク</Link>
           )}
+          <br />
+          {/*Image*/}
+          <ImageItem id={id} isHidden={result.is_rent ? true : false} />
+          {/*Rental*/}
+          <>
+            <Link to={`/rental/${id}/update`}>貸し出し情報の更新</Link>
+            <ReplaceRental id={id} />
+          </>
+          <RentRental id={id} isHidden={result.id === 1 || result.is_rent ? true : false} />
+          <br />
+          {/*Transfer*/}
+          <TransferItem id={id} isHidden={result.id === 1 || result.is_rent ? true : false} />
+          <br />
+          {/*Delete*/}
+          <DeleteItem id={id} isHidden={result.id === 1 || result.is_rent ? true : false} />
         </>
       )}
     </>
